Fix type warnings for size_t use in Firestore (#14790)

+++ Firestore/core/src/local/local_store.h
@@ -254,7 +254,7 @@ class LocalStore : public bundle::BundleCallback {
    * Runs a single backfill operation and returns the number of documents
    * processed.
    */
-  int Backfill() const;
+  size_t Backfill() const;
 
   /**
    * Returns whether the given bundle has already been loaded and its create
